Refactor codes TODO: - [ ] offboard's inheritance to planners are not transferred - [ ] test not running
