+++ car_racing/tests/ilqr_test.py
@@ -1,34 +1,35 @@
 import pickle
+
 import sympy as sp
 import numpy as np
-from racing import offboard
-from utils import base, racing_env
-from utils.constants import *
+
+from planner import *
+from racing_env import *
 
 
 def racing(args):
     track_layout = args["track_layout"]
     track_spec = np.genfromtxt("data/track_layout/" + track_layout + ".csv", delimiter=",")
     if args["simulation"]:
-        track = racing_env.ClosedTrack(track_spec, track_width=1.0)
+        track = ClosedTrack(track_spec, track_width=1.0)
         # setup ego car
-        ego = offboard.DynamicBicycleModel(name="ego", param=base.CarParam(edgecolor="black"), system_param = base.SystemParam())
+        ego = DynamicBicycleModel(name="ego", param=base.CarParam(edgecolor="black"), system_param = base.SystemParam())
         ilqr_cbf_param = base.iLQRRacingParam(vt=0.8)
         ego.set_state_curvilinear(np.zeros((X_DIM,)))
         ego.set_state_global(np.zeros((X_DIM,)))
         ego.start_logging()
-        ego.set_ctrl_policy(offboard.iLQRRacing(ilqr_cbf_param, ego.system_param))
+        ego.set_ctrl_policy(iLQRRacing(ilqr_cbf_param, ego.system_param))
         ego.ctrl_policy.set_timestep(0.1)
         ego.set_track(track)
         ego.ctrl_policy.set_track(track)
         # setup surrounding cars
         t_symbol = sp.symbols("t")
-        car1 = offboard.NoDynamicsModel(name="car1", param=base.CarParam(edgecolor="orange"))
+        car1 = NoDynamicsModel(name="car1", param=base.CarParam(edgecolor="orange"))
         car1.set_track(track)
         car1.set_state_curvilinear_func(t_symbol, 0.2 * t_symbol + 4.0, 0.1 + 0.0 * t_symbol)
         car1.start_logging()
         # setup simulation
-        simulator = offboard.CarRacingSim()
+        simulator = CarRacingSim()
         simulator.set_timestep(0.1)
         simulator.set_track(track)
         simulator.add_vehicle(ego)
