+++ car_racing/tests/overtake_planner_test.py
@@ -1,12 +1,11 @@
 import pickle
+import random
+
 import sympy as sp
 import numpy as np
-import random
-from control.lmpc_helper import LMPCPrediction
-from racing import offboard
-from utils import base, racing_env
-from utils.constants import *
 
+from planner import *
+from racing_env import *
 
 def racing_overtake(args, file_number):
     if args["save_trajectory"]:
@@ -22,7 +21,7 @@ def racing_overtake(args, file_number):
     opti_traj_xglob = np.genfromtxt(
         "data/optimal_traj/xglob_" + track_layout + ".csv", delimiter=","
     )
-    track = racing_env.ClosedTrack(track_spec, track_width=1.0)
+    track = ClosedTrack(track_spec, track_width=1.0)
     num_veh = args["number_other_agents"]
     if args["diff_alpha"]:
         alphas = [1.0, 0.9, 0.8, 0.7, 0.6, 0.5]
@@ -51,7 +50,7 @@ def racing_overtake(args, file_number):
                 ego.system_param
             )
             # define a simulator
-            simulator = offboard.CarRacingSim()
+            simulator = CarRacingSim()
             simulator.set_timestep(timestep)
             simulator.set_track(track)
             simulator.add_vehicle(ego)
@@ -244,10 +243,10 @@ def racing_overtake(args, file_number):
 
 
 def set_up_ego(timestep, track):
-    ego = offboard.DynamicBicycleModel(name="ego", param=base.CarParam(edgecolor="black"), system_param = base.SystemParam())
+    ego = DynamicBicycleModel(name="ego", param=base.CarParam(edgecolor="black"), system_param = base.SystemParam())
     ego.set_timestep(timestep)
     # run the pid controller for the first lap to collect data
-    pid_controller = offboard.PIDTracking(vt=0.7, eyt=0.0)
+    pid_controller = PIDTracking(vt=0.7, eyt=0.0)
     pid_controller.set_timestep(timestep)
     ego.set_ctrl_policy(pid_controller)
     pid_controller.set_track(track)
@@ -257,7 +256,7 @@ def set_up_ego(timestep, track):
     ego.set_track(track)
     # run mpc-lti controller for the second lap to collect data
     mpc_lti_param = base.MPCTrackingParam(vt=0.7, eyt=0.0)
-    mpc_lti_controller = offboard.MPCTracking(mpc_lti_param, ego.system_param)
+    mpc_lti_controller = MPCTracking(mpc_lti_param, ego.system_param)
     mpc_lti_controller.set_timestep(timestep)
     mpc_lti_controller.set_track(track)
     return ego, pid_controller, mpc_lti_controller
@@ -267,7 +266,7 @@ def set_up_lmpc(timestep, track, lap_number, alpha, opti_traj_xcurv, opti_traj_x
     time_lmpc = 10000 * timestep
     lmpc_param = base.LMPCRacingParam(timestep=timestep, lap_number=lap_number, time_lmpc=time_lmpc)
     racing_game_param = base.RacingGameParam(timestep=timestep, alpha=alpha, num_horizon_planner=10)
-    lmpc_controller = offboard.LMPCRacingGame(lmpc_param, racing_game_param=racing_game_param, system_param = system_param)
+    lmpc_controller = LMPCRacingGame(lmpc_param, racing_game_param=racing_game_param, system_param = system_param)
     lmpc_controller.set_track(track)
     lmpc_controller.set_timestep(timestep)
     lmpc_controller.set_opti_traj(opti_traj_xcurv, opti_traj_xglob)
@@ -280,7 +279,7 @@ def set_up_other_vehicles(track, num_veh):
     for index in range(0, num_veh):
         veh_name = "car" + str(index + 1)
         vehicles.append(
-            offboard.NoDynamicsModel(name=veh_name, param=base.CarParam(edgecolor="orange"))
+            NoDynamicsModel(name=veh_name, param=base.CarParam(edgecolor="orange"))
         )
         vehicles[index].set_track(track)
     return vehicles
