--- car_racing/utils/constants.py
@@ -1,1 +1,0 @@
-X_DIM, U_DIM = 6, 2
